pyproject.toml
[project]
name = "cortex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=7.0.0",
    "earningscall>=2.0.1",
    "edgartools>=5.30.0",
    "fastapi>=0.136.0",
    "jinja2>=3.1.6",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.45.0",
    "yfinance>=0.2.54",
]
